[bluelink] Allow float values for HV & 12V battery SoC

+++ bundles/org.openhab.binding.bluelink/src/main/java/org/openhab/binding/bluelink/internal/dto/BatteryStatus.java
@@ -19,5 +19,5 @@ import com.google.gson.annotations.SerializedName;
  *
  * @author Marcus Better - Initial contribution
  */
-public record BatteryStatus(@SerializedName("batSoc") int stateOfCharge) {
+public record BatteryStatus(@SerializedName("batSoc") double stateOfCharge) {
 }
